feat(Discussion): expose renderCommentMdast and some clean up
