--- src/components/Discussion/index.js
@@ -1,6 +1,0 @@
-export { commentComposerStorageKey, CommentComposer, CommentComposerPlaceholder } from './Composer'
-export { SecondaryAction as CommentComposerSecondaryAction } from './Internal/Composer'
-
-export { CommentList } from './Tree'
-
-export { DiscussionContext } from './DiscussionContext'
